telegram_bot/requirements.txt
telethon==1.35.0
python-dotenv==1.0.0